this is version basically works -NEW : add method to read Vfb shift of each segment at the interface and convert it to charge density -ENH : enhance the mechanism deal with difference between Window and Linux platform including different line-ending problems of the parameter file -CHG : change the name of some variables and modules

+++ tests/debug.py
@@ -3,7 +3,7 @@ import os, sys
 path = os.path.join(os.path.abspath(os.path.dirname(__file__)), os.pardir)
 if not path in sys.path:
     sys.path.append(path)
-import pytaurus.platform as platform
+import pytaurus.env as env
 import pytaurus.sentaurus as sen
 import pytaurus.sentaurus.sse as sse
 import pytaurus.sentaurus.sde as sde
@@ -11,22 +11,30 @@ import pytaurus.sentaurus.callsent as callsent
 import pytaurus.sentaurus.extract as extr
 
 def cleanDebugEnv():
-    logfile_path = os.path.join(platform.Debug_Directory, sen.Logfile_Sse)
-    if os.path.exists(logfile_path):
-        os.remove(logfile_path)
-    sentrun_path = os.path.join(platform.Debug_Directory, sen.Folder_Run_Sentaurus)
+    # clean the log file
+    logfile_paths = [os.path.join(env.Debug_Directory, sen.Logfile_Sse),
+                    os.path.join(env.Debug_Directory, sen.Logfile_Sdevice)]
+    for log_file in logfile_paths:
+        if os.path.exists(log_file):
+            os.remove(log_file)
+    #clean sentaurus running folder
+    sentrun_path = os.path.join(env.Debug_Directory, sen.Folder_Run_Sentaurus)
     for file in os.listdir(sentrun_path):
         file_path = os.path.join(sentrun_path, file)
         os.remove(file_path)
-    exchange_path = os.path.join(platform.Debug_Directory, sen.Folder_Exchange_Data)
+
+    #clean data exchage
+    exchange_path = os.path.join(env.Debug_Directory, sen.Folder_Exchange_Data)
     for file in os.listdir(exchange_path):
-        file_path = os.path.join(exchange_path, file)
-        os.remove(file_path)
+        if not file == sen.Charge_File and not file == sen.Points_Location_Subs:
+            file_path = os.path.join(exchange_path, file)
+            os.remove(file_path)
     return
 
 def test_debug():
     cleanDebugEnv()
-    debug_prj_path = platform.Debug_Directory
+    debug_prj_path = env.Debug_Directory
+    env.convertParamFile(debug_prj_path)
     trip_cells = sse.TripleCells(debug_prj_path)
     trip_cells.build()
     sse_cmd = sse.SseCmdFile(trip_cells)
